requirements.txt
streamlit>=1.35.0
torch>=2.2.0
transformers>=4.40.0
Pillow>=10.2.0
requests>=2.31.0
python-dotenv>=1.0.1


